# swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Skolaroid API
  version: 1.0.0
  description: |
    OpenAPI (Swagger) documentation for Skolaroid API route handlers under `src/app/api`.

    Notes:
    - Many successful responses include additional fields besides `success`, `message`, and `data`.
    - Some endpoints are currently placeholders or transitional implementations.
    - Auth is primarily Supabase session cookie-based unless otherwise noted.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/prisma/invitation/send:
    post:
      tags: [Invitations]
      summary: Send invitations to a group
      operationId: sendInvitations
      security:
        - SupabaseSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendInvitationsInput'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/prisma/memory/get-by-creator:
    get:
      tags: [Memories]
      summary: Get memories by creator user id
      operationId: getMemoriesByCreator
      security: []
      parameters:
        - $ref: '#/components/parameters/userIdQuery'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/prisma/memory/get-by-group:
    get:
      tags: [Memories]
      summary: Get group-only memories for a private group
      operationId: getMemoriesByGroup
      security:
        - SupabaseSession: []
      parameters:
        - $ref: '#/components/parameters/groupIdQuery'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/prisma/memory/get-by-location:
    get:
      tags: [Memories]
      summary: Get visible memories at a location
      operationId: getMemoriesByLocation
      security:
        - SupabaseSession: []
      parameters:
        - $ref: '#/components/parameters/locationIdQuery'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/prisma/memory/{memoryId}/edit:
    patch:
      tags: [Memories]
      summary: Edit memory
      operationId: editMemory
      security:
        - SupabaseSession: []
      parameters:
        - $ref: '#/components/parameters/memoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditMemoryInput'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    SupabaseSession:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase access token cookie used for authenticated user session.
    CronBearer:
      type: http
      scheme: bearer
      bearerFormat: CRON_SECRET
      description: Bearer token matching server-side `CRON_SECRET` for cron-only endpoints.
  parameters:
...
    memoryId:
      name: memoryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Memory id
...
    locationIdQuery:
      name: locationId
      in: query
      required: true
      schema:
        type: string
        format: uuid
      description: Location id
    userIdQuery:
      name: userId
      in: query
      required: true
      schema:
        type: string
        format: uuid
      description: User id
...
    groupIdQuery:
      name: groupId
      in: query
      required: true
      schema:
        type: string
        format: uuid
      description: Group id
...
  responses:
    Success:
      description: Successful operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid auth
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not authorized for this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          nullable: true
        count:
          type: integer
        deletedCount:
          type: integer
        url:
          type: string
          format: uri
      additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          nullable: true
          example: false
        error:
          type: string
        message:
          type: string
        detail:
          type: string
      additionalProperties: true
...
    SendInvitationsInput:
      type: object
      required:
        - groupId
        - emails
      properties:
        groupId:
          type: string
          format: uuid
        emails:
          type: array
          minItems: 1
          maxItems: 20
          items:
            type: string
            format: email
...
    MemoryVisibility:
      type: string
      enum: [PUBLIC, PROGRAM_ONLY, BATCH_ONLY, GROUP_ONLY, PRIVATE]
...
    EditMemoryInput:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 5000
        visibility:
          $ref: '#/components/schemas/MemoryVisibility'
        tags:
          type: array
          maxItems: 10
          items:
            type: string
            minLength: 1
            maxLength: 50
        privateGroupId:
          type: string
          format: uuid
          nullable: true
...